+++ packages/connect-multichain/src/multichain/index.ts
@@ -131,6 +131,12 @@ export class MultichainSDK extends MultichainCore {
     return this.options.storage;
   }
 
+  get transportType(): TransportType {
+    return this.__transport instanceof MWPTransport
+      ? TransportType.MWP
+      : TransportType.Browser;
+  }
+
   private get sdkInfo(): string {
     return `Sdk/Javascript SdkVersion/${getVersion()} Platform/${getPlatformType()} dApp/${this.options.dapp.url ?? this.options.dapp.name} dAppTitle/${this.options.dapp.name}`;
   }
@@ -225,7 +231,7 @@ export class MultichainSDK extends MultichainCore {
           );
           return apiTransport;
         }
-      } else if (transportType === TransportType.MPW) {
+      } else if (transportType === TransportType.MWP) {
         const { adapter: kvstore } = this.options.storage;
         const dappClient = await this.createDappClient();
         const apiTransport = new MWPTransport(dappClient, kvstore);
@@ -252,7 +258,7 @@ export class MultichainSDK extends MultichainCore {
       }
       this.state = 'connected';
       if (this.transport instanceof MWPTransport) {
-        await this.storage.setTransport(TransportType.MPW);
+        await this.storage.setTransport(TransportType.MWP);
       } else {
         await this.storage.setTransport(TransportType.Browser);
       }
@@ -312,7 +318,7 @@ export class MultichainSDK extends MultichainCore {
     this.listener = this.transport.onNotification(
       this.onTransportNotification.bind(this),
     );
-    await this.storage.setTransport(TransportType.MPW);
+    await this.storage.setTransport(TransportType.MWP);
   }
 
   private async onBeforeUnload(): Promise<void> {
@@ -383,7 +389,7 @@ export class MultichainSDK extends MultichainCore {
                 this.options.ui.factory.unload();
                 this.options.ui.factory.modal?.unmount();
                 this.state = 'connected';
-                return this.storage.setTransport(TransportType.MPW);
+                return this.storage.setTransport(TransportType.MWP);
               })
               .catch((error) => {
                 if (error instanceof ProtocolError) {
@@ -402,7 +408,7 @@ export class MultichainSDK extends MultichainCore {
         },
         async (error?: Error) => {
           if (!error) {
-            await this.storage.setTransport(TransportType.MPW);
+            await this.storage.setTransport(TransportType.MWP);
             resolve();
           } else {
             await this.storage.removeTransport();
@@ -515,6 +521,7 @@ export class MultichainSDK extends MultichainCore {
   async connect(
     scopes: Scope[],
     caipAccountIds: CaipAccountId[],
+    forceRequest?: boolean,
   ): Promise<void> {
     const { ui } = this.options;
     const platformType = getPlatformType();
@@ -527,13 +534,15 @@ export class MultichainSDK extends MultichainCore {
 
     if (this.__transport?.isConnected() && !secure) {
       return this.handleConnection(
-        this.__transport.connect({ scopes, caipAccountIds }).then(() => {
-          if (this.__transport instanceof MWPTransport) {
-            return this.storage.setTransport(TransportType.MPW);
-          } else {
-            return this.storage.setTransport(TransportType.Browser);
-          }
-        }),
+        this.__transport
+          .connect({ scopes, caipAccountIds, forceRequest })
+          .then(() => {
+            if (this.__transport instanceof MWPTransport) {
+              return this.storage.setTransport(TransportType.MWP);
+            } else {
+              return this.storage.setTransport(TransportType.Browser);
+            }
+          }),
       );
     }
 
@@ -541,7 +550,7 @@ export class MultichainSDK extends MultichainCore {
     if (platformType === PlatformType.MetaMaskMobileWebview) {
       const defaultTransport = await this.setupDefaultTransport();
       return this.handleConnection(
-        defaultTransport.connect({ scopes, caipAccountIds }),
+        defaultTransport.connect({ scopes, caipAccountIds, forceRequest }),
       );
     }
 
@@ -550,7 +559,7 @@ export class MultichainSDK extends MultichainCore {
       const defaultTransport = await this.setupDefaultTransport();
       // Web transport has no initial payload
       return this.handleConnection(
-        defaultTransport.connect({ scopes, caipAccountIds }),
+        defaultTransport.connect({ scopes, caipAccountIds, forceRequest }),
       );
     }
 
@@ -619,7 +628,11 @@ export class MultichainSDK extends MultichainCore {
         if (mobile?.preferredOpenLink) {
           mobile.preferredOpenLink(METAMASK_DEEPLINK_BASE, '_self');
         } else {
-          openDeeplink(this.options, METAMASK_DEEPLINK_BASE, METAMASK_CONNECT_BASE_URL);
+          openDeeplink(
+            this.options,
+            METAMASK_DEEPLINK_BASE,
+            METAMASK_CONNECT_BASE_URL,
+          );
         }
       }, 10); // small delay to ensure the message encryption and dispatch completes
     }